move protocol.rs to pcp.rs since pmp and pcp have compatible but different wire formats
